cdk/chat-service/requirements.txt
fastapi
uvicorn
boto3
aiobotocore
PyJWT
cryptography
httpx